requirements.txt
simplejson
mysql-connector-python